Blindly replace all insecure HTTP links with HTTPS
I know there are some broken links, but that's due to a combination
of age and not just the lack of TLS. This is a safer default, and the
way things should be moving forward. If I ever actually go through and
update the content of my site, I'll try to clean things up, but for now
it should get things happy on Netlify.