Remove Jekyll-related Gemfile
GitHub sent a security warning about the older version of Jekyll. This
setup hasn't actually worked to generate the static files for quite
a while, so I'm going to rip out the whole shebang. The migration to
another platform will be largely a manual process, but the source files
could still come in handy until that work it done.