Remove symlink setup for vault content
Due to how builds work with Cloudflare Pages, if you set a root
directory, the build command cannot access files that are symlinked
above the root. The main downside to this new structure is that the
Obsidian vault will have the name "content" locally, but that's a minor
nuisance.

Alternatively, we could change directories and then run `npm ci`
ourselves as part of the build command, but then we lose smart
build caching of the node dependencies (since they're no longer
auto-detected).